refactor(eval): extract answer_quality.rs
